Rob Willett's type tweaks for 64-bit XCode

+++ src/utstring.h
@@ -156,7 +156,7 @@ _UNUSED_ static void utstring_printf(UT_string *s, const char *fmt, ...) {
 /* Build KMP table from left to right. */
 _UNUSED_ static void _utstring_BuildTable(
     const char *P_Needle, 
-    unsigned int P_NeedleLen, 
+    size_t P_NeedleLen, 
     long *P_KMP_Table)
 {
     long i, j;
@@ -164,7 +164,7 @@ _UNUSED_ static void _utstring_BuildTable(
     i = 0;
     j = i - 1;
     P_KMP_Table[i] = j;
-    while (i < (long)P_NeedleLen)
+    while (i < P_NeedleLen)
     {
         while ( (j > -1) && (P_Needle[i] != P_Needle[j]) )
         {
@@ -172,7 +172,7 @@ _UNUSED_ static void _utstring_BuildTable(
         }
         i++;
         j++;
-        if (i < (long)P_NeedleLen)
+        if (i < P_NeedleLen)
         {
             if (P_Needle[i] == P_Needle[j])
             {
@@ -196,7 +196,7 @@ _UNUSED_ static void _utstring_BuildTable(
 /* Build KMP table from right to left. */
 _UNUSED_ static void _utstring_BuildTableR(
     const char *P_Needle, 
-    unsigned int P_NeedleLen, 
+    size_t P_NeedleLen, 
     long *P_KMP_Table)
 {
     long i, j;
@@ -206,7 +206,7 @@ _UNUSED_ static void _utstring_BuildTableR(
     P_KMP_Table[i + 1] = j;
     while (i >= 0)
     {
-        while ( (j < (long)P_NeedleLen) && (P_Needle[i] != P_Needle[j]) )
+        while ( (j < P_NeedleLen) && (P_Needle[i] != P_Needle[j]) )
         {
            j = P_KMP_Table[j + 1];
         }
@@ -241,7 +241,7 @@ _UNUSED_ static long _utstring_find(
     size_t P_NeedleLen, 
     long *P_KMP_Table)
 {
-    int i, j;
+    long i, j;
     long V_FindPosition = -1;
 
     /* Search from left to right. */
@@ -274,7 +274,7 @@ _UNUSED_ static long _utstring_findR(
     size_t P_NeedleLen, 
     long *P_KMP_Table)
 {
-    int i, j;
+    long i, j;
     long V_FindPosition = -1;
 
     /* Search from right to left. */
